requirements.txt
numpy~=1.26.4
torchtext~=0.18.0
gensim~=4.3.2
transformers~=4.42.4
torch~=2.3.1
pytorch-crf
nni~=3.0
pytorch_lightning
fasttext~=0.9.3